requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
boto3==1.40.31
botocore==1.40.31
celery==5.5.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==45.0.7
daphne==4.2.1
Django==5.2.6
django-celery-beat==2.8.1
django-cors-headers==4.8.0
django-environ==0.11.2
django-filter==24.3
django-redis==5.4.0
django-storages==1.14.6
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.1
drf-spectacular==0.28.0
gunicorn==21.2.0
python-json-logger==2.0.7
sentry-sdk==2.18.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
Pillow==10.4.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.2.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.3.1
referencing==0.36.2
rpds-py==0.27.1
s3transfer==0.14.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.10.0
zope.interface==8.0
phonenumbers==8.13.40
